requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.1.7
bidict==0.21.4
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
certifi==2020.4.5.1
cffi==1.15.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configparser==5.0.0
cryptography==36.0.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.5.1
decorator==5.1.0
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.0
duplicity==0.8.12.0
entrypoints==0.3
eventlet==0.33.0
fasteners==0.14.1
ffmpeg-python==0.2.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SocketIO==5.1.1
flask-swagger==0.2.14
flask-swagger-ui==3.25.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==0.1.16
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.2
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
log4p==2019.7.13.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
minio==5.0.10
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
MouseInfo==0.1.3
mysqlclient==2.0.1
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.6
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opencv-python-headless==4.5.2.52
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
protobuf==3.6.1
ptyprocess==0.7.0
py-dateutil==2.2
PyAutoGUI==0.9.53
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGetWindow==0.0.9
Pygments==2.10.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
pyrsistent==0.18.0
PyScreeze==0.1.28
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-engineio==4.3.1
python-socketio==5.5.1
python3-xlib==0.15
pytweening==1.0.4
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
qtconsole==5.2.1
QtPy==1.11.2
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.6.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
serial==0.0.97
simple-websocket==0.5.0
simplejson==3.16.0
six==1.15.0
SQLAlchemy==1.3.18
swagger-client==1.0.0
systemd-python==234
tacommon==0.4
terminado==0.12.1
testpath==0.5.0
tornado==6.1
traitlets==5.1.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.9
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wsproto==1.0.0
xkit==0.0.0
zipp==3.6.0
zmq==0.0.0
